# backend/internal/httpd/apispec/openapi.yaml
openapi: 3.1.0
info:
  description: Loopback-only HTTP surface served by the Go daemon. Generated from
    Go (code-first) — do not edit by hand; run `go generate ./...`.
  title: Agent Orchestrator HTTP daemon
  version: 0.1.0-route-shell
servers:
- description: Local daemon (loopback only)
  url: http://127.0.0.1:3001
paths:
...
  /api/v1/ai-accounts:
    get:
      operationId: listAIAccounts
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ControllersListAIAccountsResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: List isolated Codex and Claude account profiles
      tags:
      - settings
    post:
      operationId: createAIAccount
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ControllersCreateAIAccountRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ControllersAIAccountResponse'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Create an isolated AI account profile
      tags:
      - settings
  /api/v1/ai-accounts/{id}:
    delete:
      operationId: deleteAIAccount
      parameters:
      - description: AI account profile id.
        in: path
        name: id
        required: true
        schema:
          description: AI account profile id.
          type: string
      responses:
        "204":
          description: No Content
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Remove an AI account profile without deleting provider credentials
      tags:
      - settings
    patch:
      operationId: updateAIAccount
      parameters:
      - description: AI account profile id.
        in: path
        name: id
        required: true
        schema:
          description: AI account profile id.
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ControllersUpdateAIAccountRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ControllersAIAccountResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Update an AI account profile
      tags:
      - settings
  /api/v1/ai-accounts/{id}/credential:
    delete:
      operationId: clearAIAccountCredential
      parameters:
      - description: AI account profile id.
        in: path
        name: id
        required: true
        schema:
          description: AI account profile id.
          type: string
      responses:
        "204":
          description: No Content
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Clear only the selected AI account login credential
      tags:
      - settings
    put:
      operationId: setAIAccountCredential
      parameters:
      - description: AI account profile id.
        in: path
        name: id
        required: true
        schema:
          description: AI account profile id.
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ControllersSetAIAccountCredentialRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ControllersAIAccountResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Store a Claude profile token in the operating system credential vault
      tags:
      - settings
  /api/v1/ai-accounts/{id}/login:
    post:
      operationId: loginAIAccount
      parameters:
      - description: AI account profile id.
        in: path
        name: id
        required: true
        schema:
          description: AI account profile id.
          type: string
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ControllersAIAccountResponse'
          description: Accepted
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Internal Server Error
      summary: Start the provider-owned OAuth login for an isolated profile
      tags:
      - settings
...
components:
  schemas:
    APIError:
      properties:
        code:
          type: string
        details:
          additionalProperties: {}
          type: object
        error:
          type: string
        message:
          type: string
        requestId:
          type: string
      required:
      - error
      - code
      - message
      type: object
...
    AiaccountProfile:
      properties:
        authMethod:
          type: string
        authStatus:
          enum:
          - authenticated
          - signed_out
          - unavailable
          - authenticating
          type: string
        createdAt:
          format: date-time
          type: string
        credentialConfigured:
          type: boolean
        enabled:
          type: boolean
        harness:
          type: string
        id:
          type: string
        label:
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
      - id
      - harness
      - label
      - enabled
      - authStatus
      - credentialConfigured
      - createdAt
      - updatedAt
      type: object
...
    ControllersAIAccountResponse:
      properties:
        account:
          $ref: '#/components/schemas/AiaccountProfile'
      required:
      - account
      type: object
    ControllersCreateAIAccountRequest:
      properties:
        harness:
          enum:
          - codex
          - claude-code
          type: string
        label:
          maxLength: 40
          type: string
      required:
      - harness
      type: object
    ControllersListAIAccountsResponse:
      properties:
        accounts:
          items:
            $ref: '#/components/schemas/AiaccountProfile'
          type: array
      required:
      - accounts
      type: object
...
    ControllersSetAIAccountCredentialRequest:
      properties:
        token:
          maxLength: 3000
          minLength: 1
          type: string
      required:
      - token
      type: object
...
    ControllersUpdateAIAccountRequest:
      properties:
        enabled:
          type:
          - "null"
          - boolean
        label:
          maxLength: 40
          type:
          - "null"
          - string
      type: object
...